moved the patient component into the components folder + made css class names more clare
