+++ project/web/app/(main)/biometric-monitor/homepage.css
@@ -74,8 +74,6 @@
     gap: 30px;
   }
   
-
-  
   .right-container {
     display: flex;
     flex-direction: column;
@@ -130,16 +128,7 @@
   .bloodoxygen-percentage {
     font-size: 40px !important;
   }
-  
 
-  .test-patient{
-    display: block; /* default visible */
-  }
-  
-  .patient-dropdown-container {
-    display: none;  /* default hidden */
-
-  }
   
   .overlay {
     position: fixed;
@@ -152,20 +141,28 @@
   }
   
 
-/* Show/hide based on screen size */
+    
+  .PatientInfoContainer{
+    display: block;
+  }
+  
+  .patient-dropdown-container {
+    display: none;  
+  }
+  
 @media (max-width: 1500px) {
   .content-container {
     flex-direction: column-reverse;
   }
 
   .patient-dropdown-container {
-    display: block;  /* or flex, depending on your layout */
+    display: block;  
     display:flex;
     justify-content: center;
     align-items: center;
   }
 
-  .test-patient {
+  .PatientInfoContainer {
     display: none;
   }
 }
