+++ project/web/app/(main)/predictive-analysis/page.tsx
@@ -4,7 +4,8 @@ import HeartRateChart from "./heartRateChart";
 import BloodOxygenChart from "./bloodOxygenChart";
 
 import './predictiveAnalysis.css'
-import PatientInfo from "../biometric-monitor/PatientComponent";
+import PatientInfo from "../components/patient-component/PatientComponent";
+import PatientDropdown from "../components/patient-component/patient-dropdown";
 
 export default function PredictiveAnalysisPage() {
     return (
@@ -19,6 +20,10 @@ export default function PredictiveAnalysisPage() {
   
         <div className = "PatientInfoContainer">
           <PatientInfo/>
+        </div>
+
+        <div className = "patient-dropdown-container">
+          <PatientDropdown/>
         </div>
 
       </div>
