+++ project/web/app/(main)/predictive-analysis/predictiveAnalysis.css
@@ -31,3 +31,27 @@ html, body {
 }
   
 
+.PatientInfoContainer{
+    display: block;
+  }
+
+.patient-dropdown-container {
+    display: none;  
+}
+  
+@media (max-width: 1500px) {
+  .content-container {
+    flex-direction: column-reverse;
+  }
+
+  .patient-dropdown-container {
+    display: block;  
+    display:flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .PatientInfoContainer{
+    display: none;
+  }
+}
